Volviendo a probar que travis funcione

+++ test/run.sh
@@ -4,10 +4,15 @@ cd $(dirname $0)
 cd ../complete
 
 ./gradlew check
+ret=$?
+if [ $ret -ne 0 ]; then
+    exit $ret
+fi
+
 ./gradlew test --rerun-tasks --stacktrace
 ret=$?
 if [ $ret -ne 0 ]; then
-exit $ret
+    exit $ret
 fi
 rm -rf build
